requirements.txt
Django==3.0.4
django-debug-toolbar==2.2
django-template-toolkit==0.1.2
sentry-sdk==0.14.2
pyOpenSSL==19.1.0